Remove SetFields from Manager and Controller
