+++ pkg/source/source.go
@@ -18,28 +18,18 @@ package source
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"sync"
-	"time"
 
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/util/workqueue"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
-	logf "sigs.k8s.io/controller-runtime/pkg/internal/log"
-	"sigs.k8s.io/controller-runtime/pkg/source/internal"
+	internal "sigs.k8s.io/controller-runtime/pkg/internal/source"
 
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 )
-
-var log = logf.RuntimeLog.WithName("source")
 
 const (
 	// defaultBufferSize is the default number of event notifications that can be buffered.
@@ -69,127 +59,7 @@ type SyncingSource interface {
 
 // Kind creates a KindSource with the given cache provider.
 func Kind(cache cache.Cache, object client.Object) SyncingSource {
-	return &kind{obj: object, cache: cache}
-}
-
-// KindAsPartialMetadata modifies the given Source created with Kind(...)
-// to project the type with a metav1.PartialObjectMetadata.
-func KindAsPartialMetadata(src SyncingSource, scheme *runtime.Scheme) error {
-	ks, ok := src.(*kind)
-	if !ok {
-		return fmt.Errorf("expected Source created with source.Kind(...), got %T", src)
-	}
-
-	metaObj := &metav1.PartialObjectMetadata{}
-	gvk, err := apiutil.GVKForObject(ks.obj, scheme)
-	if err != nil {
-		return fmt.Errorf("unable to determine GVK of %T for a metadata-only watch: %w", ks.obj, err)
-	}
-	metaObj.SetGroupVersionKind(gvk)
-	ks.obj = metaObj
-	return nil
-}
-
-// kind is used to provide a source of events originating inside the cluster from Watches (e.g. Pod Create).
-type kind struct {
-	// obj is the type of object to watch.  e.g. &v1.Pod{}
-	obj client.Object
-
-	// cache used to watch APIs
-	cache cache.Cache
-
-	// started may contain an error if one was encountered during startup. If its closed and does not
-	// contain an error, startup and syncing finished.
-	started     chan error
-	startCancel func()
-}
-
-var _ SyncingSource = &kind{}
-
-// Start is internal and should be called only by the Controller to register an EventHandler with the Informer
-// to enqueue reconcile.Requests.
-func (ks *kind) Start(ctx context.Context, handler handler.EventHandler, queue workqueue.RateLimitingInterface,
-	prct ...predicate.Predicate) error {
-	if ks.obj == nil {
-		return fmt.Errorf("must create Kind with a non-nil object")
-	}
-	if ks.cache == nil {
-		return fmt.Errorf("must create Kind with a non-nil cache")
-	}
-
-	// cache.GetInformer will block until its context is cancelled if the cache was already started and it can not
-	// sync that informer (most commonly due to RBAC issues).
-	ctx, ks.startCancel = context.WithCancel(ctx)
-	ks.started = make(chan error)
-	go func() {
-		var (
-			i       cache.Informer
-			lastErr error
-		)
-
-		// Tries to get an informer until it returns true,
-		// an error or the specified context is cancelled or expired.
-		if err := wait.PollImmediateUntilWithContext(ctx, 10*time.Second, func(ctx context.Context) (bool, error) {
-			// Lookup the Informer from the Cache and add an EventHandler which populates the Queue
-			i, lastErr = ks.cache.GetInformer(ctx, ks.obj)
-			if lastErr != nil {
-				kindMatchErr := &meta.NoKindMatchError{}
-				switch {
-				case errors.As(lastErr, &kindMatchErr):
-					log.Error(lastErr, "if kind is a CRD, it should be installed before calling Start",
-						"kind", kindMatchErr.GroupKind)
-				case runtime.IsNotRegisteredError(lastErr):
-					log.Error(lastErr, "kind must be registered to the Scheme")
-				default:
-					log.Error(lastErr, "failed to get informer from cache")
-				}
-				return false, nil // Retry.
-			}
-			return true, nil
-		}); err != nil {
-			if lastErr != nil {
-				ks.started <- fmt.Errorf("failed to get informer from cache: %w", lastErr)
-				return
-			}
-			ks.started <- err
-			return
-		}
-
-		_, err := i.AddEventHandler(internal.EventHandler{Queue: queue, EventHandler: handler, Predicates: prct})
-		if err != nil {
-			ks.started <- err
-			return
-		}
-		if !ks.cache.WaitForCacheSync(ctx) {
-			// Would be great to return something more informative here
-			ks.started <- errors.New("cache did not sync")
-		}
-		close(ks.started)
-	}()
-
-	return nil
-}
-
-func (ks *kind) String() string {
-	if ks.obj != nil {
-		return fmt.Sprintf("kind source: %T", ks.obj)
-	}
-	return "kind source: unknown type"
-}
-
-// WaitForSync implements SyncingSource to allow controllers to wait with starting
-// workers until the cache is synced.
-func (ks *kind) WaitForSync(ctx context.Context) error {
-	select {
-	case err := <-ks.started:
-		return err
-	case <-ctx.Done():
-		ks.startCancel()
-		if errors.Is(ctx.Err(), context.Canceled) {
-			return nil
-		}
-		return errors.New("timed out waiting for cache to be synced")
-	}
+	return &internal.Kind{Type: object, Cache: cache}
 }
 
 var _ Source = &Channel{}
@@ -204,7 +74,7 @@ type Channel struct {
 	// Source is the source channel to fetch GenericEvents
 	Source <-chan event.GenericEvent
 
-	// Stop is to end ongoing goroutine, and close the channels
+	// stop is to end ongoing goroutine, and close the channels
 	stop <-chan struct{}
 
 	// dest is the destination channels of the added event handlers
